requirements.txt
# --- Core Model Dependencies ---
torch
transformers==4.57.1
huggingface_hub
pillow
tqdm

# --- Evaluation Metrics ---
pycocoevalcap

# --- Data Ablation & Processing ---
albumentations
opencv-python
numpy

# --- Analysis & Visualization ---
pandas
matplotlib
seaborn
ipykernel
nltk

# --- CLIPScore Dependencies ---
ftfy
regex
tqdm
clip @ git+https://github.com/openai/CLIP.git